backend/requirements.txt
fastapi
uvicorn
json_repair
sqlmodel
alembic
loguru
pydantic==2.11.7
pydantic_settings
neo4j
python-dotenv
langchain>1.0
langchain-openai
langchain-google-genai
langchain-anthropic
langchain_qwq
PyJWT